Common: comments in BQ adjustment function

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/qual/BaseQualAdjustment.java
@@ -6,7 +6,8 @@ public class BaseQualAdjustment
 
     public static final double BASE_QUAL_PERMITTED_DIFF_MIN = 0.1;
     public static final double BASE_QUAL_PERMITTED_DIFF_MAX = 2;
-    public static final double BASE_QUAL_LOWER_STEP = 1;
+
+    public static final int BASE_QUAL_LOWER_STEP = 1; // step down from the standard value if difference is within bounds
 
     public static byte adjustBaseQual(final double baseQual) { return adjustBaseQual(STANDARD_BASE_QUALS, baseQual); }
 
@@ -14,6 +15,10 @@ public class BaseQualAdjustment
     {
         if(baseQual < standardBaseQuals[1] - BASE_QUAL_PERMITTED_DIFF_MAX)
             return (byte)standardBaseQuals[0];
+
+        // for raw base qual lower than the standard value:
+        // - by the min difference or less, use the standard value, eg: 24.9 -> 25
+        // - by the max difference or less, use the standard value - step value, eg: 23.1 -> 25 - 1 = 24
 
         for(int i = standardBaseQuals.length - 1; i > 0; --i)
         {
